Fix Hibernate schema validation by aligning entity ID types with SQLite

+++ toluja-app/backend/src/main/java/com/toluja/app/dto/AuthDtos.java
@@ -5,6 +5,6 @@ import jakarta.validation.constraints.NotBlank;
 public class AuthDtos {
     public record LoginRequest(@NotBlank String username, @NotBlank String password) {}
     public record LoginResponse(String token, UserInfo user) {}
-    public record UserInfo(Long id, String username, String nomeExibicao, String role, Boolean deveTrocarSenha) {}
+    public record UserInfo(Integer id, String username, String nomeExibicao, String role, Boolean deveTrocarSenha) {}
     public record ChangePasswordRequest(@NotBlank String senhaAtual, @NotBlank String novaSenha) {}
 }
